[Add] Setting Docker/docker-compose for developments enviroment
Setted up Docker for development environment.

The docker image is used golang version 1.15.6, which is installed by default when running `brew install` in December 2020. Also used alpine, which is more lightweight.

In order to check the behavior of golang in the Docker container, stored `main.go` which prints `Hello World!`, in the `/app` directory.

With this commit, we are now able to print `Hello World!` by executing `docker-compose exec app go run main.go`.